requirements.txt
pypdf
chromadb>=1.0.9
langchain==0.3.0
langchain-google-genai==2.1.5
numpy<2
pandas==2.2.2
uvicorn[standard]==0.30.1
pydantic==2.7.4
neo4j
python-dotenv==1.0.1
fastapi==0.115.9
protobuf==3.20.3